--- lib/logincapstone/style/colorStyle.dart
@@ -1,14 +1,0 @@
-import 'package:flutter/material.dart';
-
-class ColorStyle {
-  Color primarydark = const Color(0xff101828);
-
-  Color primarblue = const Color(0xff444CE7);
-
-  Color secondarypink = const Color(0xffFFF3E2);
-
-  Color backgroundfield = const Color(0xffF9FAFB);
-
-  Color disable = const Color(0xff667085);
-  Color white = const Color(0xffFFFFFF);
-}
